file structure clean up and reorganize
