+++ frontend/src/components/avionics-Screen/AvionicsIcons.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react'
 import {Col} from 'react-bootstrap';
-import AlertDismissible from '../../AlertDismissible';
+import AlertDismissible from '../AlertDismissible';
 const AvionicsIcons = (props) => {
     const [show, setShow] = useState(false)
     return (
