+++ frontend/src/components/home-screen/Avionics.js
