+++ frontend/src/components/home-screen/Flir.js
