+++ frontend/src/components/home-screen/Mechanics.js
